@@ -3,15 +3,12 @@ import os
 import re
 import sys
 from ast import literal_eval
+from functools import lru_cache
 from os import getenv, path
 from platform import system
 from subprocess import STDOUT, CalledProcessError, check_output
 from textwrap import dedent
 
-try:
-    from functools import lru_cache
-except ImportError:
-    from backports.functools_lru_cache import lru_cache
 try:
     FileNotFoundError
 except NameError:
@@ -52,9 +49,9 @@ def check_output_u8(*args, **kwargs):
 
 def env_prefix(key, dir):
     try:
-        os.environ[key] = "%s%s%s" % (os.environ[key], os.pathsep, fspath(dir))
+        os.environ[key] = f"{os.environ[key]}{os.pathsep}{fspath(dir)}"
     except KeyError:
-        os.environ[key] = str(fspath(dir))
+        os.environ[key] = fspath(dir)
 
 
 @lru_cache()
@@ -91,14 +88,13 @@ install-matlab-engine-api-for-python-in-nondefault-locations.html
 
                 Alternatively, use `get_runtime()` instead of `get_engine()`.
                 """).format(matlabroot=matlabroot(default="matlabroot"), exe=sys.executable))
-    started = engine.find_matlab()
-    notify = False
-    if not started or (name and name not in started):
-        notify = True
-        log.debug("Starting MATLAB")
-    eng = engine.connect_matlab(name=name or getenv("SPM12_MATLAB_ENGINE", None))
-    if notify:
-        log.debug("MATLAB started")
+    log.debug("Starting MATLAB")
+    try:
+        eng = engine.connect_matlab(name=name or getenv("SPM12_MATLAB_ENGINE", None))
+    except engine.EngineError:
+        log.error("MATLAB hasn't properly cleaned up. Try restarting your computer.")
+        raise
+    log.debug("MATLAB started")
     return eng
 
 
@@ -152,7 +148,7 @@ def _install_engine():
 def get_runtime(cache="~/.mcr", version=99):
     cache = Path(cache).expanduser()
     mcr_root = cache
-    i = mcr_root / ("v%d"%version)
+    i = mcr_root / f"v{version}"
     if i.is_dir():
         mcr_root = i
     else:
@@ -174,8 +170,8 @@ def get_runtime(cache="~/.mcr", version=99):
                 if system() == "Linux":
                     install.chmod(0o755)
                     check_output_u8([
-                        fspath(install), "-P",
-                        'bean421.installLocation="%s"' % fspath(cache), "-silent"])
+                        fspath(install), "-P", f'bean421.installLocation="{fspath(cache)}"',
+                        "-silent"])
                 else:
                     raise NotImplementedError(
                         dedent("""\
@@ -185,7 +181,7 @@ def get_runtime(cache="~/.mcr", version=99):
                         """).format(fspath(install), system()))
             else:
                 raise IndexError(version)
-            mcr_root /= "v%d" % version
+            mcr_root /= f"v{version}"
             log.info("Installed")
 
     # bin
